souputils: Fix compiler warning
gstsouputils.c:35:25: error: comparison of constant 9 with expression of type
      'SoupLoggerLogLevel' is always false
      [-Werror,-Wtautological-constant-out-of-range-compare]

+++ ext/soup/gstsouputils.c
@@ -32,7 +32,7 @@ gst_soup_util_log_make_level_tag (SoupLoggerLogLevel level)
 {
   gchar c;
 
-  if (G_UNLIKELY (level > 9))
+  if (G_UNLIKELY ((gint) level > 9))
     return '?';
 
   switch (level) {
